% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadwood-package.R
\docType{package}
\name{deadwood-package}
\alias{deadwood-package}
\title{The Deadwood Outlier Detection Algorithm}
\description{
See \code{\link{deadwood}()} for more details.
}
\seealso{
The official online manual of \pkg{deadwood} at \url{https://deadwood.gagolewski.com/}
Useful links:
\itemize{
  \item \url{https://deadwood.gagolewski.com/}
  \item \url{https://github.com/gagolews/deadwood}
  \item Report bugs at \url{https://github.com/gagolews/deadwood/issues}
}

}
\author{
\strong{Maintainer}: Marek Gagolewski \email{marek@gagolewski.com} (\href{https://orcid.org/0000-0003-0637-6028}{ORCID}) [copyright holder]

}
\keyword{internal}
